// src/features/asistencia/types.ts
export type AttendanceSubsection =
  | 'no-marcaciones'
  | 'sin-credenciales'
  | 'cambios-dia'
  | 'autorizaciones';